Time: 6 ms (5.86%) | Memory: 43.4 MB (5.86%) - LeetSync

+++ 2277-count-equal-and-divisible-pairs-in-an-array/count-equal-and-divisible-pairs-in-an-array.java
@@ -12,7 +12,7 @@ class Solution {
             {
                 for(int j=i+1;j<index.size();j++)
                 {
-                    if((index.get(i)*index.get(j)) %k == 0) count++;
+                    if(((long)index.get(i)*index.get(j)) %k == 0) count++;
                 }
             }
         }
